Fixed issue where script fails because nvm is already installed

+++ utils/install-openvdm-ubuntu22.04.sh
@@ -177,16 +177,18 @@ function install_packages {
     LC_ALL=C.UTF-8 add-apt-repository -y ppa:ondrej/apache2
 
     # Install nodejs v20.11.0 LTS
-    cd ~
-    wget -qO- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.7/install.sh | bash
-    export NVM_DIR="$HOME/.nvm"
-    [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
-    [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
-    nvm install --lts
+    if [ ! -e "/usr/local/bin/npm" ]; then
+        cd ~
+        wget -qO- https://raw.githubusercontent.com/nvm-sh/nvm/v0.39.7/install.sh | bash
+        export NVM_DIR="$HOME/.nvm"
+        [ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"  # This loads nvm
+        [ -s "$NVM_DIR/bash_completion" ] && \. "$NVM_DIR/bash_completion"  # This loads nvm bash_completion
+        nvm install --lts
 
-    sudo ln -s $HOME/.nvm/versions/node/v20.11.0/bin/npm /usr/local/bin/
-    sudo ln -s $HOME/.nvm/versions/node/v20.11.0/bin/node /usr/local/bin/
-
+        sudo ln -s $HOME/.nvm/versions/node/v20.11.0/bin/npm /usr/local/bin/
+        sudo ln -s $HOME/.nvm/versions/node/v20.11.0/bin/node /usr/local/bin/
+    fi
+    
     apt-get update -qq
 
     apt install -q -y openssh-server sshpass rsync git samba smbclient \
